tests/wpt/tests/clipboard-apis/permissions-policy/clipboard-write/clipboard-write-disabled-by-permissions-policy.tentative.https.sub.html

<!doctype html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/permissions-policy/resources/permissions-policy.js"></script>
<script src="../../resources/user-activation.js"></script>
<script>
'use strict';

const same_origin_src =
  '/permissions-policy/resources/permissions-policy-clipboard-write.html';
const cross_origin_src =
  'https://{{hosts[alt][]}}:{{ports[https][0]}}' + same_origin_src;

promise_test(async t => {
  await tryGrantWritePermission();
  return promise_rejects_dom(t, 'NotAllowedError',
      navigator.clipboard.writeText('test text'));
}, 'Permissions-Policy header clipboard-write=() disallows the top-level document.');

async_test(t => {
  test_feature_availability(
    'navigator.clipboard.writeText("test text")',
    t,
    same_origin_src,
    expect_feature_unavailable_default
  );
}, 'Permissions-Policy header clipboard-write=() disallows same-origin iframes.');

async_test(t => {
  test_feature_availability(
    'navigator.clipboard.writeText("test text")',
    t,
    cross_origin_src,
    expect_feature_unavailable_default
  );
}, 'Permissions-Policy header clipboard-write=() disallows cross-origin iframes.');
</script>
</body>
